grocery space type mapping update
This was a tricky issue. The original ASHRAE supermarket model has several space types that get reduced in mapping. In the refactor it was:
 - "Produce" and "Sales" -> "grocery sales"
 - "Bakery" and "Deli" -> "food preparation"
This created an issue because the refrigeration space types would add "grocery sales" and "food preparation" refrigeration to both the corresponding original space types, because they had the same refrigeration space type. It created double equipment at the original model. Also, thermostats weren't set for the new "grocery sales" space type, which caused an EnergyPlus Error because refrigeration zones need to be conditioned.
Avoiding duplication is not necessarily a goal; refrigeration should go in a space type if it calls for it. But this was an issue created from mapping older models.

The fix is:
- make a new "bakery" space type, so "Bakery" -> "bakery" and "Deli" -> "food preparation". Only include gas equipment in the "food preparation" option. The "bakery" gets the bakery freezer/cooler, and the deli gets the kitchen freezer/cooler.
- revert "Produce" back to "retail" in the mapping, and put all refrigeration in "grocery sales". Change the space type ratios going forward to get rid of the "Produce" section and merge it with "Sales". It doesn't seem propriate to split out produce into its own space type.
- have the create typical test use the new space type ratios rather than an older model